어드민 인증해야만 페이지 접근 가능

+++ server/admin/controllers/webtoonDeleteController.js
@@ -7,13 +7,24 @@ const WebtoonDeleteController = {
         try {
             const { EnName } = req.body;
 
-            // 유효성 검사
-            if (!EnName) {
-                res.send(400, { message: '내용을 입력하세요' });
-                throw new Error('내용을 입력하세요');
-            }else{
-                const resultMessage = await WebtoonDeleteService.deleteWebtoon(EnName);
-                res.send(resultMessage); // "웹툰 삭제 성공"
+            const tokenResponse = await axios.get('http://localhost:4000/api/adminAuth', {
+                headers: {
+                Cookie: req.headers.cookie,
+                },
+            });
+            if (tokenResponse.data === '어드민 인증 성공') {
+
+                // 유효성 검사
+                if (!EnName) {
+                    res.send(400, { message: '내용을 입력하세요' });
+                    throw new Error('내용을 입력하세요');
+                }else{
+                    const resultMessage = await WebtoonDeleteService.deleteWebtoon(EnName);
+                    res.send(resultMessage); // "웹툰 삭제 성공"
+                }
+                
+            }else {
+                res.json({ message: '권한이 없습니다' });
             }
         } catch (error) {
             console.error(error);
@@ -27,13 +38,24 @@ const WebtoonDeleteController = {
         try {
             const { EnName, ep } = req.body;
 
-            // 유효성 검사
-            if (!EnName || !ep) {
-                res.send(400, { message: '내용을 입력하세요' });
-                throw new Error('내용을 입력하세요');
-            }else{
-                const resultMessage = await WebtoonDeleteService.deleteEpisode( EnName, ep);
-                res.send(resultMessage); // "에피소드 삭제 성공"
+            const tokenResponse = await axios.get('http://localhost:4000/api/adminAuth', {
+                headers: {
+                Cookie: req.headers.cookie,
+                },
+            });
+            if (tokenResponse.data === '어드민 인증 성공') {
+
+                // 유효성 검사
+                if (!EnName || !ep) {
+                    res.send(400, { message: '내용을 입력하세요' });
+                    throw new Error('내용을 입력하세요');
+                }else{
+                    const resultMessage = await WebtoonDeleteService.deleteEpisode( EnName, ep);
+                    res.send(resultMessage); // "에피소드 삭제 성공"
+                }
+                
+            }else {
+                res.json({ message: '권한이 없습니다' });
             }
         } catch (error) {
             console.error(error);
